# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IPingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponse": {
				"properties": {
					"error": {
						"properties": {
							"message": {
								"type": "string"
							},
							"code": {
								"type": "number",
								"enum": [
									401
								],
								"nullable": false
							}
						},
						"required": [
							"code"
						],
						"type": "object"
					},
					"token": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ILoginParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserResponse": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"error": {
						"properties": {
							"message": {
								"type": "string"
							},
							"code": {
								"type": "number",
								"enum": [
									409
								],
								"nullable": false
							}
						},
						"required": [
							"code"
						],
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IRegistrationParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"confirm_password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"confirm_password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPostBody": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"author": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"text",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreatePostBodyParams": {
				"properties": {
					"title": {
						"type": "string"
					},
					"text": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"text"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPatchPostsParams": {
				"properties": {
					"title": {
						"type": "string"
					},
					"text": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "blog-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/ping": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"ping"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Post",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"ping"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginParams"
							}
						}
					}
				}
			}
		},
		"/api/auth/registration": {
			"post": {
				"operationId": "Registration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/IUserResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegistrationParams"
							}
						}
					}
				}
			}
		},
		"/api/posts": {
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"page_count": {
											"type": "number",
											"format": "double"
										},
										"posts": {
											"anyOf": [
												{
													"$ref": "#/components/schemas/IPostBody"
												},
												{
													"items": {
														"$ref": "#/components/schemas/IPostBody"
													},
													"type": "array"
												}
											]
										}
									},
									"required": [
										"page_count",
										"posts"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"DESC",
								"ASC"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreatePostBodyParams"
							}
						}
					}
				}
			}
		},
		"/api/posts/{post_id}": {
			"get": {
				"operationId": "GetPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IPostBody"
										},
										{
											"items": {
												"$ref": "#/components/schemas/IPostBody"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "post_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "post_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "post_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPatchPostsParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}